world gen

+++ server/src/worldgen.rs
@@ -167,7 +167,7 @@ impl WorldGenerator for BasicWorldGenerator {
                     let biome = self.biomes.get(biome).unwrap();
                     if i == 0 {
                         for (chance, structure) in biome.get_structures() {
-                            if height / 16 == position.y && structure_rng.gen_bool(*chance as f64) {
+                            if height / 16 == position.y && structure_rng.gen_bool(*chance) {
                                 chunk.world.place_structure(
                                     BlockPosition {
                                         x: (x as i32) + (position.x * 16),
